+++ src/air/example/dummy_air.rs
@@ -3,8 +3,7 @@ use crate::{
         self,
         constraints::boundary::{BoundaryConstraint, BoundaryConstraints},
         context::AirContext,
-        trace::TraceTable,
-        AIR,
+        trace::TraceTable, air::AIR,
     },
     fri::FieldElement,
     prover::ProvingError,
