+++ src/air/example/fibonacci_rap.rs
@@ -5,8 +5,7 @@ use crate::{
         constraints::boundary::{BoundaryConstraint, BoundaryConstraints},
         context::AirContext,
         frame::Frame,
-        trace::TraceTable,
-        AIR,
+        trace::TraceTable, air::AIR,
     },
     fri::FieldElement,
     prover::ProvingError,
