+++ src/air/example/mod.rs
@@ -1,4 +1,3 @@
-pub mod cairo;
 pub mod dummy_air;
 pub mod fibonacci_2_columns;
 pub mod fibonacci_f17;
