+++ src/lib.rs
@@ -6,7 +6,7 @@ pub mod proof;
 pub mod prover;
 pub mod verifier;
 
-use air::AIR;
+use air::air::AIR;
 use lambdaworks_crypto::fiat_shamir::transcript::Transcript;
 use lambdaworks_fft::roots_of_unity::get_powers_of_primitive_root_coset;
 use lambdaworks_math::field::{
